requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch==2.1.0
openai-whisper==20231106